Move WithWpCompatOverlaySlot decorator into @wordpress/ui
The decorator previously lived in `storybook/decorators/`, but
per-component stories under `packages/ui/src/tooltip/stories/`
imported it from there — a `packages/ui` → `storybook/`
cross-package boundary violation.

Move it to `packages/ui/src/utils/with-wp-compat-overlay-slot.tsx`,
co-located with the helper it opts into. The Tooltip story now
imports it as a same-package file. The cross-library demo under
`storybook/stories/playground/` imports it via `packages/ui/src/...`
which is the normal upper-level → package import direction.

Drive-by: now imports `WP_COMPAT_OVERLAY_SLOT_ATTRIBUTE` from the
helper module instead of hardcoding the data attribute string.
